feat: New functionalities were added.
Added the while loop to allow player two to make guesses. If the guess is correct and there are remaining guesses, the player wins. If the guesses reach 10, the player loses. In addition, messages are added to inform the player of the outcome of each guess and the remaining points are updated.